+++ src/Mail/MailMessage.php
@@ -3,6 +3,7 @@
 namespace InternetGuru\LaravelCommon\Mail;
 
 use Illuminate\Notifications\Messages\MailMessage as BaseMailMessage;
+use InternetGuru\LaravelCommon\Exceptions\GeolocationServiceException;
 use InternetGuru\LaravelCommon\Services\GeolocationService;
 
 class MailMessage extends BaseMailMessage
@@ -19,7 +20,12 @@ class MailMessage extends BaseMailMessage
     public function view($view, array $data = [])
     {
         $data['ip'] = request()->ip();
-        $data['timezone'] = app(GeolocationService::class)->getLocation($data['ip'])->timezone;
+        try {
+            $timezone = app(GeolocationService::class)->getLocation($data['ip'])->timezone;
+        } catch (GeolocationServiceException $ex) {
+            $timezone = null;
+        }
+        $data['timezone'] = $timezone;
 
         return parent::view($view, $data);
     }